backend-drf/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
Django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.4.0
celery==5.4.0
redis==5.0.8
sqlalchemy==2.0.34
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.27.2
openai==1.52.0
pydantic[email]==2.9.2
requests==2.31.0
google-auth>=2.20.0
python-multipart==0.0.6
beautifulsoup4==4.12.3
faiss-cpu==1.8.0.post1
numpy==1.26.4
pandas==2.2.3
pypdf==5.0.1
sentence-transformers==3.1.1
